+++ pkg/database/database.go
@@ -10,9 +10,7 @@ import (
 	"github.com/number571/go-peer/pkg/crypto/random"
 	"github.com/number571/go-peer/pkg/crypto/symmetric"
 	"github.com/number571/go-peer/pkg/utils"
-
-	"github.com/syndtr/goleveldb/leveldb"
-	"github.com/syndtr/goleveldb/leveldb/opt"
+	"go.etcd.io/bbolt"
 )
 
 var (
@@ -20,69 +18,45 @@ var (
 )
 
 const (
-	cSaltKey  = "__SALT__"
-	cHashKey  = "__HASH__"
+	cBucket  = "_BUCKET_"
+	cSaltKey = "__SALT__"
+	cRandKey = "__RAND__"
+	cHashKey = "__HASH__"
+)
+
+const (
 	cSaltSize = 32
+	cRandSize = 32
 )
 
 type sKVDatabase struct {
 	fMutex    sync.Mutex
-	fDB       *leveldb.DB
+	fDB       *bbolt.DB
 	fSettings ISettings
 	fCipher   symmetric.ICipher
 	fAuthKey  []byte
 }
 
 func NewKVDatabase(pSett ISettings) (IKVDatabase, error) {
-	path := pSett.GetPath()
-	opt := &opt.Options{
-		DisableBlockCache: true,
-		Strict:            opt.StrictAll,
-	}
-
-	db, err := leveldb.OpenFile(path, opt)
+	db, err := bbolt.Open(pSett.GetPath(), 0600, &bbolt.Options{})
 	if err != nil {
-		openErr := utils.MergeErrors(ErrOpenDB, err)
-		db, err = tryRecover(path, opt)
-		if err != nil {
-			return nil, utils.MergeErrors(openErr, err)
-		}
+		return nil, utils.MergeErrors(ErrOpenDB, err)
 	}
 
-	isInitSalt := false
-	saltValue, err := db.Get([]byte(cSaltKey), nil)
+	saltValue, initValue, err := getSaltValue(db)
 	if err != nil {
-		if !errors.Is(err, leveldb.ErrNotFound) {
-			return nil, utils.MergeErrors(ErrReadSalt, err)
-		}
-		isInitSalt = true
-		saltValue = random.NewCSPRNG().GetBytes(cSaltSize)
-		if err := db.Put([]byte(cSaltKey), saltValue, nil); err != nil {
-			return nil, utils.MergeErrors(ErrPushSalt, err)
-		}
+		return nil, utils.MergeErrors(ErrGetSalt, err)
 	}
 
-	keyBuilder := keybuilder.NewKeyBuilder(1<<pSett.GetWorkSize(), saltValue)
-	buildKeys := keyBuilder.Build(pSett.GetPassword(), 2*symmetric.CAESKeySize)
-
-	cipherKey := buildKeys[:symmetric.CAESKeySize]
-	authKey := buildKeys[symmetric.CAESKeySize:]
-
-	if isInitSalt {
-		saltHash := hashing.NewHMACSHA256Hasher(authKey, saltValue).ToBytes()
-		if err := db.Put([]byte(cHashKey), saltHash, nil); err != nil {
-			return nil, utils.MergeErrors(ErrPushSaltHash, err)
-		}
-	}
-
-	gotSaltHash, err := db.Get([]byte(cHashKey), nil)
+	cipherKey, authKey := getCipherAuthKeys(pSett, saltValue)
+	randValue, hashRand, err := getRandHashValues(db, initValue, authKey)
 	if err != nil {
-		return nil, utils.MergeErrors(ErrReadSaltHash, err)
+		return nil, utils.MergeErrors(ErrGetHashRand, err)
 	}
 
-	newSaltHash := hashing.NewHMACSHA256Hasher(authKey, saltValue).ToBytes()
-	if !bytes.Equal(gotSaltHash, newSaltHash) {
-		return nil, ErrInvalidSaltHash
+	newHashRand := hashing.NewHMACSHA256Hasher(authKey, randValue).ToBytes()
+	if !bytes.Equal(hashRand, newHashRand) {
+		return nil, ErrInvalidHash
 	}
 
 	return &sKVDatabase{
@@ -102,10 +76,22 @@ func (p *sKVDatabase) Set(pKey []byte, pValue []byte) error {
 	defer p.fMutex.Unlock()
 
 	key := hashing.NewHMACSHA256Hasher(p.fAuthKey, pKey).ToBytes()
-	if err := p.fDB.Put(key, doEncrypt(p.fCipher, p.fAuthKey, pValue), nil); err != nil {
+	val := doEncrypt(p.fCipher, p.fAuthKey, pValue)
+
+	if err := setDB(p.fDB, key, val); err != nil {
 		return utils.MergeErrors(ErrSetValueDB, err)
 	}
 	return nil
+}
+
+func setDB(pDB *bbolt.DB, pKey []byte, pValue []byte) error {
+	return pDB.Update(func(tx *bbolt.Tx) error {
+		bucket, err := tx.CreateBucketIfNotExists([]byte(cBucket))
+		if err != nil {
+			return ErrOpenBucket
+		}
+		return bucket.Put(pKey, pValue)
+	})
 }
 
 func (p *sKVDatabase) Get(pKey []byte) ([]byte, error) {
@@ -113,7 +99,7 @@ func (p *sKVDatabase) Get(pKey []byte) ([]byte, error) {
 	defer p.fMutex.Unlock()
 
 	key := hashing.NewHMACSHA256Hasher(p.fAuthKey, pKey).ToBytes()
-	encValue, err := p.fDB.Get(key, nil)
+	encValue, err := getDB(p.fDB, key)
 	if err != nil {
 		return nil, utils.MergeErrors(ErrGetValueDB, err)
 	}
@@ -125,20 +111,43 @@ func (p *sKVDatabase) Get(pKey []byte) ([]byte, error) {
 	)
 }
 
+func getDB(pDB *bbolt.DB, pKey []byte) ([]byte, error) {
+	var encValue []byte
+	err := pDB.View(func(tx *bbolt.Tx) error {
+		b := tx.Bucket([]byte(cBucket))
+		if b == nil {
+			return ErrOpenBucket
+		}
+		val := b.Get(pKey)
+		if val == nil {
+			return ErrGetNotFound
+		}
+		encValue = make([]byte, len(val))
+		copy(encValue, val)
+		return nil
+	})
+	return encValue, err
+}
+
 func (p *sKVDatabase) Del(pKey []byte) error {
 	p.fMutex.Lock()
 	defer p.fMutex.Unlock()
 
 	key := hashing.NewHMACSHA256Hasher(p.fAuthKey, pKey).ToBytes()
-	if _, err := p.fDB.Get(key, nil); err != nil {
-		return utils.MergeErrors(ErrGetValueDB, err)
-	}
-
-	if err := p.fDB.Delete(key, nil); err != nil {
+	if err := delDB(p.fDB, key); err != nil {
 		return utils.MergeErrors(ErrDelValueDB, err)
 	}
-
 	return nil
+}
+
+func delDB(pDB *bbolt.DB, pKey []byte) error {
+	return pDB.Update(func(tx *bbolt.Tx) error {
+		b := tx.Bucket([]byte(cBucket))
+		if b == nil {
+			return ErrOpenBucket
+		}
+		return b.Delete(pKey)
+	})
 }
 
 func (p *sKVDatabase) Close() error {
@@ -151,10 +160,52 @@ func (p *sKVDatabase) Close() error {
 	return nil
 }
 
-func tryRecover(path string, opt *opt.Options) (*leveldb.DB, error) {
-	db, err := leveldb.RecoverFile(path, opt)
+func getSaltValue(pDB *bbolt.DB) ([]byte, bool, error) {
+	initValue := false
+	saltValue, err := getDB(pDB, []byte(cSaltKey))
 	if err != nil {
-		return nil, utils.MergeErrors(ErrRecoverDB, err)
+		if !errors.Is(err, ErrOpenBucket) && !errors.Is(err, ErrGetNotFound) {
+			return nil, false, utils.MergeErrors(ErrReadSalt, err)
+		}
+		initValue = true
+		saltValue = random.NewCSPRNG().GetBytes(cSaltSize)
+		if err := setDB(pDB, []byte(cSaltKey), saltValue); err != nil {
+			return nil, false, utils.MergeErrors(ErrPushSalt, err)
+		}
 	}
-	return db, nil
+	return saltValue, initValue, nil
+}
+
+func getCipherAuthKeys(pSett ISettings, pSaltValue []byte) ([]byte, []byte) {
+	keyBuilder := keybuilder.NewKeyBuilder(1<<pSett.GetWorkSize(), pSaltValue)
+	buildKeys := keyBuilder.Build(pSett.GetPassword(), 2*symmetric.CAESKeySize)
+
+	cipherKey := buildKeys[:symmetric.CAESKeySize]
+	authKey := buildKeys[symmetric.CAESKeySize:]
+
+	return cipherKey, authKey
+}
+
+func getRandHashValues(pDB *bbolt.DB, pInitValue bool, pAuthKey []byte) ([]byte, []byte, error) {
+	if pInitValue {
+		randValue := random.NewCSPRNG().GetBytes(cRandSize)
+		if err := setDB(pDB, []byte(cRandKey), randValue); err != nil {
+			return nil, nil, utils.MergeErrors(ErrPushRand, err)
+		}
+		hashRand := hashing.NewHMACSHA256Hasher(pAuthKey, randValue).ToBytes()
+		if err := setDB(pDB, []byte(cHashKey), hashRand); err != nil {
+			return nil, nil, utils.MergeErrors(ErrPushHashRand, err)
+		}
+		return randValue, hashRand, nil
+	}
+
+	randValue, err := getDB(pDB, []byte(cRandKey))
+	if err != nil {
+		return nil, nil, utils.MergeErrors(ErrReadRand, err)
+	}
+	hashRand, err := getDB(pDB, []byte(cHashKey))
+	if err != nil {
+		return nil, nil, utils.MergeErrors(ErrReadHashRand, err)
+	}
+	return randValue, hashRand, nil
 }